add user entities in domain layer
